Simplify. No error handling needed for table.all()
I like the table['title', 'text'] syntax, but it always returns an
OrderedDict sorted alphabetically by the first index (title). Worse, it
barfs if the table is empty. If I just go back to .all() or .find(),
it's simpler, no errors, and in ID (creation) order.
Probably these complications are a good argument for using a proper
schema instead of letting dataset hand everything.  But it works out
okay for educational purposes.